drive-backend/requirements.txt
aiosmtplib==5.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
Authlib==1.6.5
bcrypt==4.0.1
blinker==1.9.0
boto3==1.42.54
botocore==1.42.54
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dnspython==2.8.0
dotenv==0.9.9
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.131.0
fastapi-mail==1.6.2
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
jmespath==1.1.0
MarkupSafe==3.0.3
motor==3.7.1
passlib==1.7.4
pillow==12.1.1
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pymongo==4.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
razorpay==2.0.1
regex==2025.11.3
rsa==4.9.1
s3transfer==0.16.0
six==1.17.0
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.33.0
watchfiles==1.1.1
websockets==16.0
httpx==0.28.1
itsdangerous==2.2.0
zipfile36==0.1.3
